requirements.txt
fastapi
uvicorn[standard]
python-dotenv
PyPDF2
langchain-google-genai
python-multipart

